perf: Directly read gbq table for simple plans
